module3/mwe/github-actions-docker/requirements.txt
# GitHub Actions Docker MWE - Minimal dependencies
# Only PyYAML is needed to parse workflow files

PyYAML==6.0.1